refactor: organize CEF sources
