+++ src/cef/subprocess/main.cpp
@@ -2,7 +2,7 @@
 #include "include/cef_render_process_handler.h"
 #include "include/cef_command_line.h"
 #include "include/wrapper/cef_message_router.h"
-#include "../framework/stdext/format.h"
+#include "../../framework/stdext/format.h"
 
 #ifdef _WIN32
 #  include <windows.h>
