+++ src/framework/luafunctions.cpp
@@ -50,7 +50,7 @@
 #include <framework/ui/ui.h>
 #include <framework/ui/uiwebview.h>
 #ifdef USE_CEF
-#include <framework/ui/uicefwebview.h>
+#include <cef/ui/uicefwebview.h>
 #endif
 #include <framework/input/mouse.h>
 #endif
